+++ src/hooks/useEditableFormElementIds.ts
@@ -1,0 +1,9 @@
+import * as React from 'react'
+
+export const EditableFormElementIdsContext = React.createContext<
+  string[] | undefined
+>(undefined)
+
+export default function useEditableFormElementIds(): string[] | undefined {
+  return React.useContext(EditableFormElementIdsContext)
+}